test/requirements.txt
numpy==1.24.4
paho-mqtt==1.6.1
